Правда или действие (4) - разбил form на отдельные модули для фото и текста, добавил валидаторы на каждую ошибку тега
